backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pypdf2
pillow
python-dotenv
requests
